finished error handling exercise.
Errors6.rs is not a well-designed exercise as there are a lot of error
types with mutual conversions but such chaos is overwhelming to
learners. Letting alone that this `puzzle` coule be solved with a single
`map_error` appended at s.parse().

Exercise should be not only straightforward but also educational woth
neither achived by error6.