finished merge list
the list implementation is using NonNull type and it is complicated and
confusing, we have simpler ways of doing this. Not to mention the unsafe
operations are needed to manupulate the list.
Also, the types adopted are messy and in need of multiple manual
conversion.
Last, there are a lot of places which variables do not need to be
mutable.